# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "rtfd 是自托管的 Sphinx 文档构建与托管服务，本文档描述其 HTTP API（版本号与 assets/VERSION 保持一致）。\n所有接口挂 /rtfd 前缀，响应统一为 {\"success\":bool,\"message\":string,\"data\":any}，\n业务错误同样以该结构返回（HTTP 200），由 customHTTPErrorHandler 统一处理。\n鉴权：管理接口（项目管理、配置查询、导入导出）与项目级接口（详情/更新/删除/导出）采用 HMAC-SHA256 动态签名，\n需随请求携带 X-Rtfd-Ts（Unix 秒）、X-Rtfd-Nonce（随机串）、X-Rtfd-Sign（HMAC-SHA256 签名）三个头；\n签名串 = ts + \"\\n\" + nonce（以 secret 为 HMAC 密钥，与 method/path/body 无关），密钥为 [api] secret 或项目 secret（二者其一即可）。\n可用 `rtfd sign` 生成三个头，并在 Swagger UI 右上角 Authorize 中分别填入 X-Rtfd-Ts / X-Rtfd-Nonce / X-Rtfd-Sign；构建与 webhook 接口使用项目 secret（HMAC/Token）。\n多数接口兼容两种路径风格：/rtfd/{name}/xxx 与 /rtfd/xxx/{name}（文档只列出前者）。",
        "title": "Rtfd API",
        "contact": {},
        "license": {
            "name": "Apache-2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "2.0.0"
    },
    "basePath": "/",
    "paths": {
        "/rtfd/github/app": {
            "post": {
                "description": "接收 installation / installation_repositories 事件（需 header X-GitHub-Hook-Installation-Target-Type=integration），\n校验 App ID 后同步项目 meta 与仓库级 webhook",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhook"
                ],
                "summary": "GitHub App 事件回调",
                "responses": {
                    "200": {
                        "description": "处理成功",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    },
                    "default": {
                        "description": "事件类型、目标类型或 App ID 校验失败",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    }
                }
            }
        },
...
        "/rtfd/{name}/badge": {
            "get": {
                "description": "返回 SVG 状态徽章（passing/failing/unknown），可直接内嵌到 README。无需鉴权。\n路径别名：GET /rtfd/badge/{name}",
                "produces": [
                    "image/svg+xml"
                ],
                "tags": [
                    "挂件"
                ],
                "summary": "文档状态徽章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分支或tag，缺省用项目最新版本；latest 亦表示最新版本",
                        "name": "branch",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SVG 内容",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "查询构建结果失败",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    }
                }
            }
        },
        "/rtfd/{name}/build": {
            "post": {
                "security": [
                    {
                        "RtfdNonce": [],
                        "RtfdSign": [],
                        "RtfdTs": []
                    }
                ],
                "description": "异步触发指定项目的文档构建，接口立即返回 201，构建结果可经项目详情（build=1）查询。\n路径别名：POST /rtfd/build/{name}；鉴权：项目 secret（未设置则免鉴权）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "构建"
                ],
                "summary": "触发构建",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分支或tag，缺省用项目最新版本",
                        "name": "branch",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "为 true/on/1 时输出完整构建日志（BuildWithAll）",
                        "name": "debug",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "已异步启动，branch 回显实际分支",
                        "schema": {
                            "$ref": "#/definitions/api.resb"
                        }
                    },
                    "default": {
                        "description": "项目不存在或签名校验失败",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    }
                }
            }
        },
        "/rtfd/{name}/desc": {
            "get": {
                "description": "返回项目 URL、语言、最新版本、可用版本、构建器、是否隐藏git等元数据，供文档页面浮动挂件（rtfd.js）渲染。\n无需鉴权。路径别名：GET /rtfd/desc/{name}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "挂件"
                ],
                "summary": "获取项目描述",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名称（不区分大小写）",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "项目元数据，data.versions 为 {语言: [版本]} 映射",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.resd"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.descData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "项目不存在或数据目录无效（success=false）",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    }
                }
            }
        },
...
        "/rtfd/{name}/webhook": {
            "post": {
                "description": "接收 GitHub（校验 X-Hub-Signature HMAC-SHA1）与 Gitee（校验 X-Gitee-Token）的 push/release 事件，\n校验通过后异步构建；ping 事件直接返回 pong；命中项目 meta 的 excluded_branch 时忽略。\n路径别名：POST /rtfd/webhook/{name}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhook"
                ],
                "summary": "git仓库webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ping 事件回 pong（success=true，ping=pong）",
                        "schema": {
                            "$ref": "#/definitions/api.resp"
                        }
                    },
                    "201": {
                        "description": "已异步启动构建（命中 excluded_branch 时返回 200 且 success=false）",
                        "schema": {
                            "$ref": "#/definitions/api.resb"
                        }
                    },
                    "default": {
                        "description": "来源、事件类型或签名校验失败",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.descData": {
            "type": "object",
            "properties": {
                "builder": {
                    "description": "Sphinx 构建器",
                    "type": "string"
                },
                "defaultBranch": {
                    "description": "系统默认分支",
                    "type": "string"
                },
                "dn": {
                    "description": "自定义域名，未设置时为 false"
                },
                "gsp": {
                    "description": "git 服务商（GitHub/Gitee）",
                    "type": "string"
                },
                "hideGit": {
                    "description": "是否隐藏git入口（非html构建器时恒为true）",
                    "type": "boolean"
                },
                "lang": {
                    "description": "语言列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "latest": {
                    "description": "最新版本（分支或tag）",
                    "type": "string"
                },
                "public": {
                    "description": "是否公开仓库",
                    "type": "boolean"
                },
                "showNav": {
                    "description": "是否显示挂件导航",
                    "type": "boolean"
                },
                "single": {
                    "description": "是否单一版本",
                    "type": "boolean"
                },
                "sourceDir": {
                    "description": "文档源目录",
                    "type": "string"
                },
                "url": {
                    "description": "仓库地址（私有仓库会剥敏）",
                    "type": "string"
                },
                "versions": {
                    "description": "各语言的可用版本：{语言: [版本]}",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "api.res": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.resb": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.resd": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.resp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "ping": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "RtfdNonce": {
            "description": "动态签名随机串，须与 X-Rtfd-Ts、X-Rtfd-Sign 三者同时携带",
            "type": "apiKey",
            "name": "X-Rtfd-Nonce",
            "in": "header"
        },
        "RtfdSign": {
            "description": "动态签名 HMAC-SHA256(secret, ts+\"\\n\"+nonce)，须与 X-Rtfd-Ts、X-Rtfd-Nonce 三者同时携带",
            "type": "apiKey",
            "name": "X-Rtfd-Sign",
            "in": "header"
        },
        "RtfdTs": {
            "description": "动态签名时间戳（Unix 秒），须与 X-Rtfd-Nonce、X-Rtfd-Sign 三者同时携带",
            "type": "apiKey",
            "name": "X-Rtfd-Ts",
            "in": "header"
        }
    }
}